fix(ci): unbreak develop — dedupe i18next + clear latent test/arch failures
Root cause of the red develop CI: two i18next versions (26.3.1 and 26.3.2,
both satisfying `^26.0.0`) were resolved in the tree, so react-i18next was
duplicated. Component `useTranslation` and the test `I18nextProvider` bound to
different physical copies → split React context → components rendered raw i18n
keys, failing ~127 render tests across the react-ui-* packages. `pnpm dedupe`
collapses i18next to a single version.

Also sweeps the failures accumulated on develop:
- arch readme-presence: add @granit/react-icons README
- arch no-empty-catch: log via createLogger in 3 localStorage-write catches
  (react-documents, react-ui-blob-storage, react-ui-kit + new package logger)
- arch no-undeclared-dep: declare stories deps in react-ui-ai-prompts
- lint import-order: react-ui-account passkeys-page
- orphan tests left by the #799/#800 refactors: relocate the useDebouncedValue
  and parseUserAgent tests to their canonical packages, move the formatCurrency
  test to @granit/utils, drop the removed usePagination test, and fix the stale
  group-detail-page not-found assertion (EmptyState migration)

lint ✓ · tsc ✓ · arch-tests ✓ · 11207 tests pass
