+++ packages/@granit/react-ui-account/src/security/passkeys-page.tsx
@@ -25,8 +25,8 @@ import {
   Spinner,
   toast,
 } from '@granit/react-ui';
-import { EmptyState } from '@granit/react-ui-kit';
 import { fromBase64Url } from '@granit/react-ui-authentication-local';
+import { EmptyState } from '@granit/react-ui-kit';
 import { Fingerprint, Pencil, Plus, Trash2 } from 'lucide-react';
 import { useState } from 'react';
