+++ packages/@granit/react-ui-invoicing/src/__tests__/modules.test.ts
@@ -1,5 +1,4 @@
 import { INVOICE_WORKFLOW_STATES } from '../constants';
-import { formatCurrency } from '../lib/format-currency';
 import { logger } from '../logger';
 
 describe('INVOICE_WORKFLOW_STATES', () => {
@@ -18,22 +17,5 @@ describe('logger', () => {
   it('should expose a logger instance with the standard log methods', () => {
     expect(typeof logger.error).toBe('function');
     expect(typeof logger.info).toBe('function');
-  });
-});
-
-describe('formatCurrency', () => {
-  it('should format minor units as a locale-aware currency string', () => {
-    expect(formatCurrency(12100, 'EUR', 'en')).toBe('€121.00');
-  });
-
-  it('should respect the requested locale for separators', () => {
-    // French formatting uses a comma decimal separator and a trailing symbol.
-    const formatted = formatCurrency(123456, 'EUR', 'fr-FR');
-    expect(formatted).toContain('234,56');
-    expect(formatted).toContain('€');
-  });
-
-  it('should fall back to the runtime default locale when none is given', () => {
-    expect(formatCurrency(5000, 'USD')).toContain('50');
   });
 });
