+++ packages/@granit/react-ui-kit/package.json
@@ -15,6 +15,7 @@
   },
   "peerDependencies": {
     "@granit/data-lookup": "workspace:*",
+    "@granit/logger": "workspace:*",
     "@granit/query-engine": "workspace:*",
     "@granit/react-api-client": "workspace:*",
     "@granit/react-data-lookup": "workspace:*",
@@ -43,6 +44,7 @@
   "devDependencies": {
     "@granit/api-client": "workspace:*",
     "@granit/data-lookup": "workspace:*",
+    "@granit/logger": "workspace:*",
     "@granit/query-engine": "workspace:*",
     "@granit/react-api-client": "workspace:*",
     "@granit/react-data-lookup": "workspace:*",
