+++ packages/@granit/react-ui-kit/src/logger.ts
@@ -1,0 +1,4 @@
+import { createLogger } from '@granit/logger';
+
+/** Package-scoped logger for cross-cutting UI-kit utilities. */
+export const logger = createLogger('react-ui-kit');
